+++ lib/rspec/core/formatters/base_formatter.rb
@@ -42,7 +42,7 @@ module RSpec
         # This will only be invoked once, and the next one to be invoked
         # is {#example_group_started}.
         #
-        # @param notification [NullNotification]
+        # @param notification [StartNotification]
         def start(notification)
           start_sync_output
           @example_count = notification.count