+++ spec/rspec/core/formatters/progress_formatter_spec.rb
@@ -5,7 +5,7 @@ RSpec.describe RSpec::Core::Formatters::ProgressFormatter do
   include FormatterSupport
 
   before do
-    send_notification :start, count_notification(2)
+    send_notification :start, start_notification(2)
     allow(formatter).to receive(:color_enabled?).and_return(false)
   end
 
@@ -25,7 +25,7 @@ RSpec.describe RSpec::Core::Formatters::ProgressFormatter do
   end
 
   it "produces standard summary without pending when pending has a 0 count" do
-    send_notification :dump_summary, summary_notification(0.00001, 2, 0, 0)
+    send_notification :dump_summary, summary_notification(0.00001, 2, 0, 0, 0)
     expect(output.string).to match(/^\n/)
     expect(output.string).to match(/2 examples, 0 failures/i)
     expect(output.string).not_to match(/0 pending/i)